+++ src/vue/components/flow.vue
@@ -1,12 +1,13 @@
 <template>
 	<div class="main">
 		<padding-bar></padding-bar>
-		<div v-if="imgTopper" class="title">
-			<img :src="imgTopper">
+		<div v-if="imgSetting" class="title">
+			<img v-if="imgSetting.mode=='image'" :src="imgSetting.topper">
+			<span v-if="imgSetting.mode=='text'" v-html="imgSetting.topper"></span>
 		</div>
 		<div class="content" @click="goOut">
 			<div>
-				<qrcode :content="content" v-if="mode=='qrcode'" :imgCenter="imgCenter"></qrcode>
+				<qrcode :content="content" v-if="mode=='qrcode'" :imgSetting="imgSetting"></qrcode>
 			</div>
 			<div class="bottom">
 				<slot></slot>
@@ -26,10 +27,9 @@
 		computed: {
 
 		},
-		mounted: function () {
-			console.log(this.imgTopper);
-		},
-		props: ['mode', 'content', 'imgTopper', 'imgCenter'],
+		watch: {},
+		mounted: function () {},
+		props: ['mode', 'content', 'imgSetting'],
 		components: {
 			"qrcode": qrcode,
 			"padding-bar": paddingbar