+++ src/vue/components/qrcode.vue
@@ -22,7 +22,10 @@
                 </tr>
             </table>
         </div>
-        <div class="above"><img v-if="imgCenter" :src="imgCenter"></div>
+        <div class="above" v-if="imgSetting">
+            <img class="imgs" v-if="imgSetting.mode=='image'" :src="imgSetting.center">
+            <span class="imgs" v-if="imgSetting.mode=='text'" v-html="imgSetting.center"></span>
+        </div>
     </div>
 </template>
 
@@ -33,7 +36,9 @@
     export default {
         data: function () {
             return {
-                src: ""
+                src: "",
+                image: false,
+                text: false
             }
         },
         computed: {
@@ -42,12 +47,15 @@
         watch: {
             content: function () {
                 this.src = qrtable(this.content, 6);
+            },
+            imgSetting: function () {
+                this.src = qrtable(this.content, 6);
             }
         },
         mounted: function () {
             this.src = qrtable(this.content, 6);
         },
-        props: ['content', 'imgCenter'],
+        props: ['content', 'imgSetting'],
         components: {},
         methods: {
             toClass: function (ingo) {
@@ -69,7 +77,7 @@
         border: 0px;
     }
 
-    img {
+    .imgs {
         width: 48px;
         height: 48px;
         background-color: rgba(216, 117, 0, 0.95);